Changed type of pixel priority so we can use negative values.

+++ src/graphics/pixel.hpp
@@ -8,20 +8,20 @@ class Pixel
 {
   public:
     Pixel() = default;
-    Pixel(byte colorId, Palette* palette, byte spritePriority, byte backgroundPriority);
+    Pixel(byte colorId, Palette* palette, int spritePriority, int backgroundPriority);
     ~Pixel() = default;
     byte getColorId() const;
     Palette* getPalette() const;
-    byte getSpritePriority() const;
-    byte getBackgroundPriority() const;
+    int getSpritePriority() const;
+    int getBackgroundPriority() const;
     bool operator==(const Pixel& rhs) const;
     bool operator!=(const Pixel& rhs) const;
 
   private:
     byte _colorId;
     Palette* _palette;
-    byte _spritePriority;
-    byte _backgroundPriority;
+    int _spritePriority;
+    int _backgroundPriority;
 };
 
 #endif // GROUBOY_PIXEL_HPP
